Refactor variable declaration handling in typechecker and main.go, and add TwoDArray type
