+++ language/function.wal
@@ -46,7 +46,7 @@ fn testFn(param1: int, param2: FN) -> int {
     ret param2(param1, 5.7);
 }
 
-testFn(4, fn (a: int, b?: float = 1.0) -> int {
+testFn(4, fn (a: int, b?: float = 10.0) -> int {
     ret a + b;
 });
