+++ main.go
@@ -13,7 +13,7 @@ import (
 
 func main() {
 	fmt.Println("Hello world!")
-	filePath := "language/structs.wal"
+	filePath := "language/arrayAccess.wal"
 	tokens := lexer.Tokenize(filePath, false)
 	parser := parseMachine.NewParser(filePath, tokens)
 	tree := parser.Parse()
