+++ parser/expr.go
@@ -120,222 +120,6 @@ func parseGroupingExpr(p *Parser) ast.Node {
 	return expr
 }
 
-// parseVarAssignmentExpr parses a variable assignment expression in the source code.
-// It takes a parser instance, the left-hand side node, and the binding power as arguments.
-// The function ensures that the left-hand side of the assignment is a valid identifier,
-// array index access, or struct property access. If not, it generates an error message.
-// It then advances the parser to the operator, parses the right-hand side expression,
-// and constructs a VarAssignmentExpr node with the appropriate location information.
-//
-// Parameters:
-//   - p: *Parser - The parser instance.
-//   - left: ast.Node - The left-hand side node of the assignment.
-//   - bp: BINDING_POWER - The binding power for the expression.
-//
-// Returns:
-//   - ast.Node - The parsed variable assignment expression node.
-func parseVarAssignmentExpr(p *Parser, left ast.Node, bp BINDING_POWER) ast.Node {
-
-	switch left.(type) {
-	case ast.IdentifierExpr:
-		break
-	case ast.ArrayIndexAccess:
-		break
-	case ast.StructPropertyAccessExpr:
-		break
-	default:
-		errMsg := "cannot assign to a non-identifier\n"
-		errgen.MakeError(p.FilePath, left.StartPos().Line, left.EndPos().Line, left.StartPos().Column, left.EndPos().Column, errMsg).Display()
-	}
-
-	operator := p.advance()
-
-	right := parseExpr(p, bp)
-
-	endPos := right.EndPos()
-
-	return ast.VarAssignmentExpr{
-		Assignee: left,
-		Value:    right,
-		Operator: operator,
-		Location: ast.Location{
-			Start: left.StartPos(),
-			End:   endPos,
-		},
-	}
-}
-
-// parseArrayExpr parses an array expression from the input tokens.
-// It expects the current token to be the opening bracket '[' of the array.
-// The function will consume tokens until it reaches the closing bracket ']'.
-// Elements within the array are expected to be separated by commas.
-// Returns an ast.ArrayExpr node representing the parsed array expression.
-//
-// Parameters:
-//   - p: A pointer to the Parser instance.
-//
-// Returns:
-//   - ast.Node: An AST node representing the parsed array expression.
-func parseArrayExpr(p *Parser) ast.Node {
-
-	start := p.advance().Start //eat the [ token
-
-	var values []ast.Node
-
-	for p.currentTokenKind() != lexer.CLOSE_BRACKET {
-		value := parseExpr(p, PRIMARY_BP)
-		values = append(values, value)
-		if p.currentTokenKind() != lexer.CLOSE_BRACKET {
-			p.expect(lexer.COMMA_TOKEN)
-		}
-	}
-
-	end := p.expect(lexer.CLOSE_BRACKET).End
-
-	return ast.ArrayLiteral{
-		Values: values,
-		Location: ast.Location{
-			Start: start,
-			End:   end,
-		},
-	}
-}
-
-// parseArrayAccess parses an array access expression from the input.
-// It expects the current token to be an opening bracket '[' and parses
-// the index expression inside the brackets. The function returns an
-// ast.ArrayIndexAccess node representing the array access.
-//
-// Parameters:
-// - p: A pointer to the Parser instance.
-// - left: The left-hand side node, representing the array being accessed.
-// - bp: The binding power for the expression parsing.
-//
-// Returns:
-// - An ast.Node representing the array index access.
-func parseArrayAccess(p *Parser, left ast.Node, bp BINDING_POWER) ast.Node {
-	p.expect(lexer.OPEN_BRACKET)
-	index := parseExpr(p, bp)
-	end := p.expect(lexer.CLOSE_BRACKET).End
-	return ast.ArrayIndexAccess{
-		Arrayvalue: left,
-		Index:      index,
-		Location: ast.Location{
-			Start: left.StartPos(),
-			End:   end,
-		},
-	}
-}
-
-// parseStructLiteral parses a struct literal from the input tokens.
-// It expects the following sequence of tokens:
-// - An '@' token indicating the start of a struct literal.
-// - An identifier token representing the struct name.
-// - An opening curly brace '{'.
-// - A series of property definitions, each consisting of:
-//   - An identifier token for the property name.
-//   - A colon ':' token.
-//   - An expression representing the property value.
-//   - An optional comma ',' token if there are more properties.
-//
-// - A closing curly brace '}'.
-//
-// The function returns an ast.Node representing the parsed struct literal.
-//
-// Parameters:
-// - p: A pointer to the Parser instance.
-//
-// Returns:
-// - An ast.Node representing the parsed struct literal.
-func parseStructLiteral(p *Parser) ast.Node {
-
-	start := p.expect(lexer.AT_TOKEN).Start
-
-	idetifierToken := p.expectError(lexer.IDENTIFIER_TOKEN, fmt.Errorf("expected a struct name"))
-
-	identidier := ast.IdentifierExpr{
-		Name: idetifierToken.Value,
-		Location: ast.Location{
-			Start: idetifierToken.Start,
-			End:   idetifierToken.End,
-		},
-	}
-
-	p.expect(lexer.OPEN_CURLY)
-
-	//parse the values
-	props := map[string]ast.Node{}
-
-	for p.hasToken() && p.currentTokenKind() != lexer.CLOSE_CURLY {
-		//we expect an identifier
-		iden := p.expect(lexer.IDENTIFIER_TOKEN)
-		//then we expect colon
-		p.expect(lexer.COLON_TOKEN)
-		//now we expect value as expression
-		val := parseExpr(p, DEFAULT_BP)
-
-		if _, ok := props[iden.Value]; ok {
-			errgen.MakeError(p.FilePath, iden.Start.Line, iden.End.Line, iden.Start.Column, iden.End.Column, fmt.Sprintf("property '%s' was previously assigned", iden.Value)).Display()
-		}
-
-		props[iden.Value] = val
-
-		//if the next token is not } then we have more values
-		if p.currentTokenKind() != lexer.CLOSE_CURLY {
-			//we expect comma
-			p.expect(lexer.COMMA_TOKEN)
-		}
-	}
-
-	end := p.expect(lexer.CLOSE_CURLY).End
-
-	structVal := ast.StructLiteral{
-		Identifier: identidier,
-		Properties: props,
-		Location: ast.Location{
-			Start: start,
-			End:   end,
-		},
-	}
-
-	return structVal
-}
-
-// parsePropertyExpr parses a property access expression from the parser.
-// It expects the current token to be a dot (.) followed by an identifier.
-// The function constructs and returns an AST node representing the property access.
-//
-// Parameters:
-// - p: A pointer to the Parser instance.
-// - left: The left-hand side node of the property access expression.
-// - bp: The binding power (precedence) of the expression.
-//
-// Returns:
-// - An AST node representing the property access expression.
-func parsePropertyExpr(p *Parser, left ast.Node, bp BINDING_POWER) ast.Node {
-
-	p.expect(lexer.DOT_TOKEN)
-
-	identifier := p.expect(lexer.IDENTIFIER_TOKEN)
-
-	property := ast.IdentifierExpr{
-		Name: identifier.Value,
-		Location: ast.Location{
-			Start: identifier.Start,
-			End:   identifier.End,
-		},
-	}
-
-	return ast.StructPropertyAccessExpr{
-		Object:   left,
-		Property: property,
-		Location: ast.Location{
-			Start: left.StartPos(),
-			End:   property.End,
-		},
-	}
-}
-
 // parsePostfixExpr parses a postfix expression, which consists of an identifier
 // followed by a postfix operator (e.g., increment or decrement).
 //
@@ -458,44 +242,6 @@ func parseBinaryExpr(p *Parser, left ast.Node, bp BINDING_POWER) ast.Node {
 		Location: ast.Location{
 			Start: left.StartPos(),
 			End:   right.EndPos(),
-		},
-	}
-}
-
-// parseCallExpr parses a function call expression in the source code.
-// It expects the current token to be an open parenthesis and will parse
-// all arguments until it encounters a closing parenthesis.
-//
-// Parameters:
-// - p: A pointer to the Parser instance.
-// - left: The left-hand side node, typically the function being called.
-// - bp: The binding power, which determines the precedence of the expression.
-//
-// Returns:
-//   - An ast.Node representing the function call expression, including the
-//     caller, arguments, and their location in the source code.
-func parseCallExpr(p *Parser, left ast.Node, bp BINDING_POWER) ast.Node {
-
-	p.advance() //eat the open paren
-	startPos := left.StartPos()
-	var args []ast.Node
-	// parse the arguments
-	for p.currentTokenKind() != lexer.CLOSE_PAREN {
-		arg := parseExpr(p, DEFAULT_BP)
-		args = append(args, arg)
-		if p.currentTokenKind() != lexer.CLOSE_PAREN {
-			p.expect(lexer.COMMA_TOKEN)
-		}
-	}
-
-	endPos := p.expect(lexer.CLOSE_PAREN).End
-
-	return ast.FunctionCallExpr{
-		Caller:    left,
-		Arguments: args,
-		Location: ast.Location{
-			Start: startPos,
-			End:   endPos,
 		},
 	}
 }
