+++ parser/stmt.go
@@ -1,7 +1,6 @@
 package parser
 
 import (
-	"errors"
 	"fmt"
 	"strings"
 	"walrus/ast"
@@ -9,85 +8,6 @@ import (
 	"walrus/lexer"
 )
 
-// parseVarDeclStmt parses a variable declaration statement in the source code.
-// It handles both `let` and `const` declarations, with optional type annotations
-// and initial values.
-//
-// Parameters:
-// - p: A pointer to the Parser instance.
-//
-// Returns:
-// - An ast.Node representing the parsed variable declaration statement.
-//
-// The function performs the following steps:
-// 1. Advances the parser to consume the `let` or `const` keyword.
-// 2. Determines if the declaration is a constant.
-// 3. Expects and consumes an identifier token for the variable name.
-// 4. Optionally parses an explicit type if a colon `:` is present.
-// 5. Parses the assignment operator `:=` or `=` and the initial value expression if present.
-// 6. Ensures that constants have an initial value.
-// 7. Expects and consumes a semicolon `;` to terminate the statement.
-// 8. Constructs and returns an ast.VarDeclStmt node with the parsed information.
-func parseVarDeclStmt(p *Parser) ast.Node {
-
-	declToken := p.advance() // advance the let/const keyword
-
-	// is it let or const?
-	isConst := declToken.Kind == lexer.CONST_TOKEN
-
-	// parse the variable name
-	identifier := p.expect(lexer.IDENTIFIER_TOKEN)
-
-	// parse the explicit type if present. This will be nil if no type is specified.
-	var explicitType ast.DataType
-
-	var value ast.Node
-
-	assignmentToken := p.advance()
-
-	if assignmentToken.Kind == lexer.COLON_TOKEN {
-		// syntax is let a : <type>
-		explicitType = parseType(p, DEFAULT_BP)
-	} else if assignmentToken.Kind != lexer.WALRUS_TOKEN {
-		msg := "Invalid variable declaration syntax"
-		errgen.MakeError(p.FilePath, assignmentToken.Start.Line, assignmentToken.End.Line, assignmentToken.Start.Column, assignmentToken.End.Column, msg).AddHint("Maybe you want to use : or := instead of =", errgen.TEXT_HINT).Display()
-	}
-
-	if p.currentTokenKind() != lexer.SEMI_COLON_TOKEN {
-		// then we have an assignment
-		if assignmentToken.Kind == lexer.COLON_TOKEN {
-			p.expect(lexer.EQUALS_TOKEN)
-		}
-		value = parseExpr(p, ASSIGNMENT_BP)
-	}
-
-	//if const, we must have a value
-	if isConst && value == nil {
-		msg := "constants must have value when declared"
-		errgen.MakeError(p.FilePath, p.currentToken().Start.Line, p.currentToken().End.Line, p.currentToken().Start.Column, p.currentToken().End.Column, msg).Display()
-	}
-
-	end := p.expectSemicolon().End
-
-	node := ast.VarDeclStmt{
-		Variable: ast.IdentifierExpr{
-			Name: identifier.Value,
-			Location: ast.Location{
-				Start: identifier.Start,
-				End:   identifier.End,
-			},
-		},
-		Value:        value,
-		ExplicitType: explicitType,
-		IsConst:      isConst,
-		Location: ast.Location{
-			Start: declToken.Start,
-			End:   end,
-		},
-	}
-
-	return node
-}
 
 // parseUserDefinedTypeStmt parses a user-defined type statement in the source code.
 // It expects the 'type' keyword followed by an identifier that starts with a capital letter.
@@ -112,7 +32,7 @@ func parseUserDefinedTypeStmt(p *Parser) ast.Node {
 
 	udType := parseUDTType(p)
 
-	p.expectSemicolon()
+	p.expect(lexer.SEMI_COLON_TOKEN)
 
 	return ast.TypeDeclStmt{
 		UDType:     udType,
@@ -155,356 +75,6 @@ func parseBlock(p *Parser) ast.BlockStmt {
 	}
 }
 
-// parseForStmt parses a 'for' statement in the source code.
-// It handles different types of 'for' loops:
-// - Infinite loop: `for { }`
-// - Condition-only loop: `for i < 10 { }`
-// - Traditional for loop with initialization, condition, and increment: `for i := 0; i < 10; i++ { }`
-// - For-each loop: `for v in arr { }` or `for i, v in arr { }`
-//
-// Parameters:
-// - p: A pointer to the Parser instance.
-//
-// Returns:
-// - An ast.Node representing the parsed 'for' statement.
-// - If the syntax is invalid, it generates an error and returns nil.
-func parseForStmt(p *Parser) ast.Node {
-	start := p.advance().Start // eat for token
-	//first token is either an identifier or open curly
-	if p.currentTokenKind() == lexer.OPEN_CURLY {
-		// infinite loop
-		block := parseBlock(p)
-		return ast.ForStmt{
-			Start:     nil,
-			Condition: nil,
-			Increment: nil,
-			Block:     block,
-			Location: ast.Location{
-				Start: start,
-				End:   block.End,
-			},
-		}
-	} else if p.currentTokenKind() == lexer.IDENTIFIER_TOKEN {
-
-		identifier := p.currentToken()
-
-		idententifierExpr := ast.IdentifierExpr{
-			Name: identifier.Value,
-			Location: ast.Location{
-				Start: identifier.Start,
-				End:   identifier.End,
-			},
-		}
-
-		switch p.nextTokenKind() {
-		case lexer.WALRUS_TOKEN: // for i := 0; i < 10; i++ { }
-			p.advance() // eat the identifier
-			p.advance() // eat the walrus token
-			// value of the identifier
-			value := parseExpr(p, ASSIGNMENT_BP)
-			p.expect(lexer.SEMI_COLON_TOKEN)
-			// condition
-			condition := parseExpr(p, ASSIGNMENT_BP)
-			p.expect(lexer.SEMI_COLON_TOKEN)
-			// increment
-			increment := parseExpr(p, ASSIGNMENT_BP)
-
-			block := parseBlock(p)
-
-			return ast.ForStmt{
-				Start: ast.VarDeclStmt{
-					Variable:     idententifierExpr,
-					Value:        value,
-					ExplicitType: nil,
-					IsConst:      false,
-					Location: ast.Location{
-						Start: identifier.Start,
-						End:   value.EndPos(),
-					},
-				},
-				Condition: condition,
-				Increment: increment,
-				Block:     block,
-				Location: ast.Location{
-					Start: start,
-					End:   block.End,
-				},
-			}
-		case lexer.IN_TOKEN: // for v in arr { }
-			p.advance() // eat the identifier
-			p.advance() // eat the in token
-			// value of the identifier
-			value := parseExpr(p, ASSIGNMENT_BP)
-
-			block := parseBlock(p)
-
-			return ast.ForEachStmt{
-				Key:      nil,
-				Value:    idententifierExpr,
-				Iterable: value,
-				Block:    block,
-				Location: ast.Location{
-					Start: start,
-					End:   block.End,
-				},
-			}
-		case lexer.COMMA_TOKEN: // for i, v in arr { }
-			p.advance() // eat the identifier
-			p.advance() // eat the comma token
-			// value of the identifier
-			key := idententifierExpr
-			value := ast.IdentifierExpr{
-				Name: p.expect(lexer.IDENTIFIER_TOKEN).Value,
-				Location: ast.Location{
-					Start: identifier.Start,
-					End:   identifier.End,
-				},
-			}
-
-			p.expect(lexer.IN_TOKEN)
-
-			// value of the identifier
-			iterable := parseExpr(p, ASSIGNMENT_BP)
-
-			block := parseBlock(p)
-
-			return ast.ForEachStmt{
-				Key:      key,
-				Value:    value,
-				Iterable: iterable,
-				Block:    block,
-				Location: ast.Location{
-					Start: start,
-					End:   block.End,
-				},
-			}
-
-		default:
-			// for i < 10 { } // condition only
-			condition := parseExpr(p, ASSIGNMENT_BP)
-			block := parseBlock(p)
-			return ast.ForStmt{
-				Start: ast.IdentifierExpr{
-					Name: identifier.Value,
-					Location: ast.Location{
-						Start: identifier.Start,
-						End:   identifier.End,
-					},
-				},
-				Condition: condition,
-				Increment: nil,
-				Block:     block,
-				Location: ast.Location{
-					Start: start,
-					End:   block.End,
-				},
-			}
-		}
-
-	} else {
-		//error
-		//msg := "invalid for loop syntax"
-		errgen.MakeError(p.FilePath, start.Line, start.Line, start.Column, start.Column, "invalid for loop syntax").Display()
-		return nil
-	}
-}
-
-// parseIfStmt parses an if statement from the input and returns an AST node representing the if statement.
-// It expects the parser to be positioned at the 'if' token at the start of the if statement.
-//
-// The function performs the following steps:
-// 1. Advances the parser to consume the 'if' token and records the start position.
-// 2. Parses the condition expression of the if statement.
-// 3. Parses the consequent block of the if statement.
-// 4. Checks for the presence of an 'else' token and, if found, parses the alternate block or another if statement.
-//
-// The returned AST node includes the condition, the consequent block, and optionally the alternate block.
-//
-// Parameters:
-// - p: A pointer to the Parser instance.
-//
-// Returns:
-// - An AST node representing the parsed if statement.
-func parseIfStmt(p *Parser) ast.Node {
-
-	start := p.advance().Start // eat if token
-
-	condition := parseExpr(p, ASSIGNMENT_BP)
-
-	//parse block
-	consequentBlock := parseBlock(p)
-
-	var alternate ast.Node
-
-	if p.hasToken() && p.currentTokenKind() == lexer.ELSE_TOKEN {
-		p.advance() // eat else token
-		if p.hasToken() && p.currentTokenKind() == lexer.IF_TOKEN {
-			alternate = parseIfStmt(p)
-		} else {
-			alternate = parseBlock(p)
-		}
-	}
-
-	return ast.IfStmt{
-		Condition:      condition,
-		Block:          consequentBlock,
-		AlternateBlock: alternate,
-		Location: ast.Location{
-			Start: start,
-			End:   consequentBlock.End,
-		},
-	}
-}
-
-// parseLambdaFunction parses a lambda function expression from the input and returns an AST node representing the function.
-// It expects the parser to be positioned at the start of the lambda function.
-//
-// The function performs the following steps:
-// 1. Advances the parser to consume the 'fn' token and records the start position.
-// 2. Parses the function signature, including parameters and return type.
-// 3. Parses the function body block.
-//
-// Returns:
-// - ast.Node: An AST node representing the parsed lambda function expression, including its parameters, return type, body, and location.
-func parseLambdaFunction(p *Parser) ast.Node {
-	//annonymous function
-	start := p.advance().Start // eat fn token
-
-	params, returnType := parseFunctionSignature(p)
-
-	block := parseBlock(p)
-
-	return ast.FunctionLiteral{
-		Params:     params,
-		ReturnType: returnType,
-		Body:       block,
-		Location: ast.Location{
-			Start: start,
-			End:   block.End,
-		},
-	}
-}
-
-// parseFunctionDeclStmt parses a function declaration statement from the input
-// and returns an AST node representing the function declaration.
-//
-// The function expects the parser to be positioned at the start of the function
-// declaration (i.e., the 'fn' token). It advances the parser, consumes the
-// function name, parses the function signature (parameters and return type),
-// and then parses the function body block.
-//
-// Parameters:
-//   - p: A pointer to the Parser instance.
-//
-// Returns:
-//   - ast.Node: An AST node representing the function declaration, which includes
-//     the function's identifier, parameters, return type, and body block.
-func parseFunctionDeclStmt(p *Parser) ast.Node {
-
-	start := p.advance().Start // eat fn token
-
-	nameToken := p.expect(lexer.IDENTIFIER_TOKEN)
-
-	params, returnType := parseFunctionSignature(p)
-
-	block := parseBlock(p)
-
-	return ast.FunctionDeclStmt{
-		Identifier: ast.IdentifierExpr{
-			Name: nameToken.Value,
-			Location: ast.Location{
-				Start: nameToken.Start,
-				End:   nameToken.End,
-			},
-		},
-		FunctionLiteral: ast.FunctionLiteral{
-			Params:     params,
-			ReturnType: returnType,
-			Body:       block,
-			Location: ast.Location{
-				Start: start,
-				End:   block.End,
-			},
-		},
-	}
-}
-
-// parseFunctionSignature parses the signature of a function, including its parameters and return type.
-// It expects the function signature to start with an opening parenthesis and end with a closing parenthesis.
-// Parameters can be either required or optional, and optional parameters must have a default value.
-// The return type is optional and, if present, is indicated by an arrow (->) followed by the type.
-//
-// Parameters:
-// - p (*Parser): The parser instance used to parse the function signature.
-//
-// Returns:
-// - ([]ast.FunctionParam): A slice of FunctionParam representing the parameters of the function.
-// - (ast.DataType): The return type of the function.
-func parseFunctionSignature(p *Parser) ([]ast.FunctionParam, ast.DataType) {
-	p.expect(lexer.OPEN_PAREN)
-
-	//parse params
-	var params []ast.FunctionParam
-
-	for p.hasToken() && p.currentTokenKind() != lexer.CLOSE_PAREN {
-		paramToken := p.expect(lexer.IDENTIFIER_TOKEN)
-		param := ast.IdentifierExpr{
-			Name: paramToken.Value,
-			Location: ast.Location{
-				Start: paramToken.Start,
-				End:   paramToken.End,
-			},
-		}
-		// if : then the param is not optional
-		// if ?: then the param is optional
-		currentToken := p.currentToken()
-
-		if currentToken.Kind != lexer.COLON_TOKEN && currentToken.Kind != lexer.OPTIONAL_TOKEN {
-			errgen.MakeError(p.FilePath, currentToken.Start.Line, currentToken.End.Line, currentToken.Start.Column, currentToken.End.Column, "expected : or ?:").Display()
-		}
-
-		p.advance()
-
-		isOptional := currentToken.Kind == lexer.OPTIONAL_TOKEN
-
-		paramType := parseType(p, DEFAULT_BP)
-
-		var defaultValue ast.Node
-
-		if isOptional {
-			p.expectError(lexer.EQUALS_TOKEN, errors.New("expected default value for optional parameter. eg: param?: int = 10"))
-			defaultValue = parseExpr(p, ASSIGNMENT_BP)
-		}
-
-		params = append(params, ast.FunctionParam{
-			Identifier:   param,
-			Type:         paramType,
-			IsOptional:   isOptional,
-			DefaultValue: defaultValue,
-			Location: ast.Location{
-				Start: param.Start,
-				End:   paramType.EndPos(),
-			},
-		})
-
-		if p.currentTokenKind() != lexer.CLOSE_PAREN {
-			p.expect(lexer.COMMA_TOKEN)
-		}
-	}
-
-	p.expect(lexer.CLOSE_PAREN)
-
-	var returnType ast.DataType
-
-	//parse return type which is optional
-	if p.currentTokenKind() != lexer.OPEN_CURLY {
-		p.expect(lexer.ARROW_TOKEN)
-		returnType = parseType(p, DEFAULT_BP)
-	}
-
-	return params, returnType
-}
-
 // parseReturnStmt parses a return statement in the source code.
 // It expects the current token to be a return token and advances the parser.
 // If the next token is not a semicolon, it parses an expression for the return value.
@@ -525,148 +95,10 @@ func parseReturnStmt(p *Parser) ast.Node {
 		value = parseExpr(p, ASSIGNMENT_BP)
 	}
 
-	end := p.expectSemicolon().End
+	end := 	p.expect(lexer.SEMI_COLON_TOKEN).End
 
 	return ast.ReturnStmt{
 		Value: value,
-		Location: ast.Location{
-			Start: start,
-			End:   end,
-		},
-	}
-}
-
-func parseInterfaceStmt(p *Parser) ast.Node {
-	start := p.advance().Start // eat impl token
-
-	interfaceName := p.expect(lexer.IDENTIFIER_TOKEN)
-
-	p.expect(lexer.OPEN_CURLY)
-
-	methods := make(map[string]ast.InterfaceMethod)
-
-	for p.hasToken() && p.currentTokenKind() != lexer.CLOSE_CURLY {
-
-		start := p.expect(lexer.FUNCTION).Start
-
-		if p.currentTokenKind() != lexer.IDENTIFIER_TOKEN {
-			errgen.MakeError(p.FilePath, p.currentToken().Start.Line, p.currentToken().End.Line, p.currentToken().Start.Column, p.currentToken().End.Column, "expected method name").Display()
-		}
-
-		name := p.expect(lexer.IDENTIFIER_TOKEN)
-
-		dataType, params, returnType := getFunctionTypeSignature(p)
-
-		if _, ok := methods[name.Value]; ok {
-			msg := fmt.Sprintf("method %s already defined", name.Value)
-			errgen.MakeError(p.FilePath, name.Start.Line, name.End.Line, name.Start.Column, name.End.Column, msg).Display()
-		}
-
-		methods[name.Value] = ast.InterfaceMethod{
-			Identifier: ast.IdentifierExpr{
-				Name: name.Value,
-				Location: ast.Location{
-					Start: name.Start,
-					End:   name.End,
-				},
-			},
-			FunctionType: ast.FunctionType{
-				TypeName:   dataType,
-				Parameters: params,
-				ReturnType: returnType,
-				Location: ast.Location{
-					Start: start,
-					End:   returnType.EndPos(),
-				},
-			},
-		}
-
-		if p.currentTokenKind() != lexer.CLOSE_CURLY {
-			p.expect(lexer.SEMI_COLON_TOKEN)
-		}
-	}
-
-	end := p.expect(lexer.CLOSE_CURLY).End
-
-	return ast.InterfaceDeclStmt{
-		Interface: ast.IdentifierExpr{
-			Name: interfaceName.Value,
-			Location: ast.Location{
-				Start: interfaceName.Start,
-				End:   interfaceName.End,
-			},
-		},
-		Methods: methods,
-		Location: ast.Location{
-			Start: start,
-			End:   end,
-		},
-	}
-}
-
-func parseImplementStmt(p *Parser) ast.Node {
-
-	start := p.advance().Start // eat impl token
-
-	typeName := p.expect(lexer.IDENTIFIER_TOKEN)
-
-	p.expect(lexer.OPEN_CURLY)
-
-	methods := make(map[string]ast.ImplMethod)
-
-	for p.hasToken() && p.currentTokenKind() != lexer.CLOSE_CURLY {
-
-		IsPrivate := false
-
-		if p.currentTokenKind() == lexer.PRIVATE_TOKEN {
-			IsPrivate = true
-			p.advance()
-		}	
-
-		p.expect(lexer.FUNCTION)
-
-		fnName := p.expect(lexer.IDENTIFIER_TOKEN)
-
-		params, ret := parseFunctionSignature(p)	
-
-		body := parseBlock(p)
-		
-		method := ast.ImplMethod{
-			IsPrivate: IsPrivate,
-			FunctionDeclStmt: ast.FunctionDeclStmt{
-				Identifier: ast.IdentifierExpr{
-					Name: fnName.Value,
-					Location: ast.Location{
-						Start: fnName.Start,
-						End:   fnName.End,
-					},
-				},
-				FunctionLiteral: ast.FunctionLiteral{
-					Params:    params,
-					ReturnType: ret,
-					Body: 	   body,
-					Location: ast.Location{
-						Start: fnName.Start,
-						End:   body.End,
-					},
-				},
-			},
-		}
-
-		methods[fnName.Value] = method
-	}
-
-	end := p.expect(lexer.CLOSE_CURLY).End
-
-	return ast.ImplStmt{
-		ImplFor: ast.IdentifierExpr{
-			Name: typeName.Value,
-			Location: ast.Location{
-				Start: typeName.Start,
-				End:   typeName.End,
-			},
-		},
-		Methods: methods,
 		Location: ast.Location{
 			Start: start,
 			End:   end,
