+++ typechecker/identifiers.go
@@ -15,6 +15,10 @@ func checkIdentifier(node ast.IdentifierExpr, env *TypeEnvironment) ValueTypeInt
 	}
 	// if we found value on that scope, return the value. Else make error (though there is no change to reach the error)
 	if val, ok := scope.variables[name]; ok {
+		val, err := getValueTypeInterface(val, env)
+		if err != nil {
+			errgen.MakeError(env.filePath, node.StartPos().Line, node.EndPos().Line, node.StartPos().Column, node.EndPos().Column, err.Error()).Display()
+		}
 		return val
 	}
 	errgen.MakeError(env.filePath, node.StartPos().Line, node.EndPos().Line, node.StartPos().Column, node.EndPos().Column, "failed to check type. not found in environment").Display()
