+++ typechecker/typedecl.go
@@ -1,7 +1,9 @@
 package typechecker
 
 import (
+	"fmt"
 	"walrus/ast"
+	"walrus/builtins"
 	"walrus/errgen"
 )
 
@@ -9,67 +11,47 @@ func checkTypeDeclaration(node ast.TypeDeclStmt, env *TypeEnvironment) ValueType
 
 	typeName := node.UDType
 
+	fmt.Printf("declaring type %s\n", node.UDTypeName)
+
 	var val ValueTypeInterface
 
 	switch t := typeName.(type) {
 	case ast.StructType:
-		props := map[string]StructProperty{}
-		for propname, propval := range t.Properties {
-			propType := EvaluateTypeName(propval.PropType, env)
-			property := StructProperty{
-				IsPrivate: propval.IsPrivate,
-				Type:      propType,
-			}
-			props[propname] = property
-		}
-
-		val = Struct{
-			DataType:   STRUCT_TYPE,
-			StructName: node.UDTypeName,
-			Elements:   props,
-			Methods:   	map[string]StructMethod{},
-		}
+		val = checkStructTypeDecl(node.UDTypeName, t, env)
 	case ast.FunctionType:
-		val = checkFunctionSignature(node.UDTypeName, t, env)
+		val = EvaluateTypeName(t, env)
+	case ast.InterfaceType:
+		val = checkInterfaceTypeDecl(node.UDTypeName, t, env)
+	case ast.ArrayType:
+		val = EvaluateTypeName(t.ArrayType, env)
+		arr := Array{
+			DataType:  builtins.ARRAY,
+			ArrayType: val,
+		}
+		val = arr
+	case nil:
+		val = Void{
+			DataType: VOID_TYPE,
+		}
 	default:
-		typ := EvaluateTypeName(node.UDType, env)
+		typ, err := stringToValueTypeInterface(VALUE_TYPE(t.Type()), env)
+		if err != nil {
+			errgen.MakeError(env.filePath, t.StartPos().Line, t.EndPos().Line, t.StartPos().Column, t.EndPos().Column, err.Error()).Display()
+		}
 		val = typ
 	}
 
 	typeVal := UserDefined{
 		DataType: USER_DEFINED_TYPE,
+		TypeName: node.UDTypeName,
 		TypeDef:  val,
 	}
 
 	err := env.DeclareType(node.UDTypeName, typeVal)
 	if err != nil {
+		fmt.Printf("cannot declare type %s: %s\n", node.UDTypeName, err.Error())
 		errgen.MakeError(env.filePath, node.Start.Line, node.End.Line, node.Start.Column, node.End.Column, err.Error()).Display()
 	}
 
-	return nil
-}
-
-func checkFunctionSignature(name string, method ast.FunctionType, env *TypeEnvironment) Fn {
-	var params []FnParam
-
-	funcEnv := NewTypeENV(env, FUNCTION_SCOPE, name, env.filePath)
-
-	for _, param := range method.Parameters {
-		typ := EvaluateTypeName(param.Type, funcEnv)
-
-		params = append(params, FnParam{
-			Name:       param.Identifier.Name,
-			IsOptional: param.IsOptional,
-			Type:       typ,
-		})
-	}
-
-	ret := EvaluateTypeName(method.ReturnType, funcEnv)
-
-	return Fn{
-		DataType:      FUNCTION_TYPE,
-		Params:        params,
-		Returns:       ret,
-		FunctionScope: *funcEnv,
-	}
+	return val
 }
