+++ typechecker/utils.go
@@ -25,74 +25,30 @@ func RandStringRunes(n int) string {
 }
 
 // generate interfaces from the type enum
-func makeTypesInterface(typ VALUE_TYPE, env *TypeEnvironment) (ValueTypeInterface, error) {
-	switch typ {
-	case INT_TYPE:
-		return Int{
-			DataType: typ,
-		}, nil
-	case FLOAT_TYPE:
-		return Float{
-			DataType: typ,
-		}, nil
-	case CHAR_TYPE:
-		return Chr{
-			DataType: typ,
-		}, nil
-	case STRING_TYPE:
-		return Str{
-			DataType: typ,
-		}, nil
-	case BOOLEAN_TYPE:
-		return Bool{
-			DataType: typ,
-		}, nil
-	case NULL_TYPE:
-		return Null{
-			DataType: typ,
-		}, nil
-	case VOID_TYPE:
-		return Void{
-			DataType: typ,
-		}, nil
-	default:
-		//search for the type
-		typeEnv, err := env.ResolveType(string(typ))
-		if err != nil {
-			return nil, err
-		}
+func stringToValueTypeInterface(typ VALUE_TYPE, env *TypeEnvironment) (ValueTypeInterface, error) {
 
-		return UserDefined{
-			DataType: 	typ,
-			TypeDef: 	typeEnv.types[string(typ)],
-		}, nil
+	builtin, ok := env.builtins[string(typ)]
+	if ok {
+		return builtin, nil
 	}
+
+	//search for the type
+	typeEnv, err := env.ResolveType(string(typ))
+	if err != nil {
+		return nil, err
+	}
+
+	return typeEnv.types[string(typ)].(UserDefined).TypeDef, nil
 }
 
-func handleExplicitType(explicitType ast.DataType, env *TypeEnvironment) ValueTypeInterface {
-	//Explicit type is defined
-	var expectedTypeInterface ValueTypeInterface
-	switch t := explicitType.(type) {
-	case ast.ArrayType:
-		val := EvaluateTypeName(t, env)
-		expectedTypeInterface = val
-	default:
-		val, err := makeTypesInterface(VALUE_TYPE(explicitType.Type()), env)
-		if err != nil {
-			errgen.MakeError(env.filePath, explicitType.StartPos().Line, explicitType.EndPos().Line, explicitType.StartPos().Column, explicitType.EndPos().Column, err.Error()).Display()
-		}
-		expectedTypeInterface = val
-	}
-	return expectedTypeInterface
-}
-
-
-func getTypename(typeName ValueTypeInterface) VALUE_TYPE {
+func valueTypeInterfaceToString(typeName ValueTypeInterface) VALUE_TYPE {
 	switch t := typeName.(type) {
 	case Array:
-		return "[]" + getTypename(t.ArrayType)
+		return "[]" + valueTypeInterfaceToString(t.ArrayType)
 	case Struct:
 		return VALUE_TYPE(t.StructName)
+	case Interface:
+		return VALUE_TYPE(t.InterfaceName)
 	case Fn:
 		ParamStrs := ""
 		for i, param := range t.Params {
@@ -102,30 +58,40 @@ func getTypename(typeName ValueTypeInterface) VALUE_TYPE {
 			} else {
 				ParamStrs += ": "
 			}
-			ParamStrs += string(getTypename(param.Type))
+			ParamStrs += string(valueTypeInterfaceToString(param.Type))
 			if i != len(t.Params)-1 {
 				ParamStrs += ", "
 			}
 		}
-		ReturnStr := string(getTypename(t.Returns))
+		ReturnStr := string(valueTypeInterfaceToString(t.Returns))
 		if ReturnStr != "" {
 			ReturnStr = " -> " + ReturnStr
 		}
 		return VALUE_TYPE(fmt.Sprintf("fn(%s)%s", ParamStrs, ReturnStr))
 	case UserDefined:
-		return getTypename(t.TypeDef)
+		return valueTypeInterfaceToString(t.TypeDef)
 	default:
+		fmt.Printf("Default case %T's value %v\n", t, t.DType())
 		return t.DType()
 	}
 }
 
+
 func MatchTypes(expected, provided ValueTypeInterface, filePath string, lineStart, lineEnd, colStart, colEnd int) {
 
-	expectedType := getTypename(expected)
-	gotType := getTypename(provided)
+	expectedType := valueTypeInterfaceToString(expected)
+	gotType := valueTypeInterfaceToString(provided)
+
+	fmt.Printf("Expected: %s, Got: %s\n", expectedType, gotType)
+	fmt.Printf("Expected: %T, Got: %T\n", expected, provided)
+
 
 	if expectedType != gotType {
-		errgen.MakeError(filePath, lineStart, lineEnd, colStart, colEnd, fmt.Sprintf("typecheck:expected '%s', got '%s'", expectedType, gotType)).Display()
+		if expected.DType() == INTERFACE_TYPE {
+			checkMethodsImplementations(expected, provided, filePath, lineStart, lineEnd, colStart, colEnd)
+			return
+		}
+		errgen.MakeError(filePath, lineStart, lineEnd, colStart, colEnd, fmt.Sprintf("typecheck:cannot assign type '%s' to type '%s'", gotType, expectedType)).Display()
 	}
 }
 
@@ -162,10 +128,10 @@ func IsNumberType(operand ValueTypeInterface) bool {
 //   - A ValueTypeInterface representing the evaluated type.
 //
 // The function performs the following steps:
-//   1. If the dtype is an ArrayType, it recursively evaluates the element type and returns an Array.
-//   2. If the dtype is a FunctionType, it evaluates the parameter types and return type, creates a new function scope, and returns a Fn.
-//   3. If the dtype is nil, it returns a Void type.
-//   4. For other types, it attempts to create a ValueTypeInterface and handles any errors that occur.
+//  1. If the dtype is an ArrayType, it recursively evaluates the element type and returns an Array.
+//  2. If the dtype is a FunctionType, it evaluates the parameter types and return type, creates a new function scope, and returns a Fn.
+//  3. If the dtype is nil, it returns a Void type.
+//  4. For other types, it attempts to create a ValueTypeInterface and handles any errors that occur.
 func EvaluateTypeName(dtype ast.DataType, env *TypeEnvironment) ValueTypeInterface {
 	switch t := dtype.(type) {
 	case ast.ArrayType:
@@ -176,14 +142,13 @@ func EvaluateTypeName(dtype ast.DataType, env *TypeEnvironment) ValueTypeInterfa
 		}
 		return arr
 	case ast.FunctionType:
-
 		var params []FnParam
-
 		for _, param := range t.Parameters {
 			paramType := EvaluateTypeName(param.Type, env)
 			params = append(params, FnParam{
-				Name: param.Identifier.Name,
-				Type: paramType,
+				Name: 		param.Identifier.Name,
+				IsOptional: param.IsOptional,
+				Type: 		paramType,
 			})
 		}
 
@@ -197,13 +162,13 @@ func EvaluateTypeName(dtype ast.DataType, env *TypeEnvironment) ValueTypeInterfa
 			Returns:       returns,
 			FunctionScope: *scope,
 		}
-
 	case nil:
 		return Void{
 			DataType: VOID_TYPE,
 		}
 	default:
-		val, err := makeTypesInterface(VALUE_TYPE(t.Type()), env)
+		fmt.Printf("Default case value %v\n", t.Type())
+		val, err := stringToValueTypeInterface(VALUE_TYPE(t.Type()), env)
 		if err != nil {
 			errgen.MakeError(env.filePath, dtype.StartPos().Line, dtype.EndPos().Line, dtype.StartPos().Column, dtype.EndPos().Column, err.Error()).Display()
 		}
